refactor: modify the rule to be more real-world representation
